backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
google-genai==1.5.0
python-dotenv==1.0.1
